LayoutTests/platform/glib/imported/w3c/web-platform-tests/css/cssom/getComputedStyle-insets-relpos-inline-expected.txt

Lorem ipsum dolor sit amet
Lorem ipsum dolor sit amet

FAIL OOF with left fixed right auto in relpos inline container assert_equals: expected "140px" but got "127px"
FAIL OOF with left auto right fixed in relpos inline container assert_equals: expected "140px" but got "3px"
FAIL OOF with left fixed right auto in relpos inline container with mixed directions assert_equals: expected "140px" but got "127px"
FAIL OOF with left auto right fixed in relpos inline container with mixed directions assert_equals: expected "140px" but got "3px"
